refactor: restructure project files and update component paths
feat: add new components and update existing ones
style: improve CSS styling and layout consistency
chore: reorganize project directory structure
