--- _Components/Catagory/Catagory.module.css
@@ -1,20 +1,0 @@
-.catagory{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-    margin: 20px 0;
-}
-.catagoryItem{
-    width: 30%;
-    height: 100px;
-    background-color: #f1f1f1;
-    margin: 10px 0;
-}
-.imgContiner{
-    position: relative;
-    width: 100%;
-    height: 100%;
-    
-}
