--- app/news-list/[category]/page.module.css
@@ -1,66 +1,0 @@
-.sidebar {
-  width: 20%;
-  height: 100%;
-  background-color: #f1f1f1;
-  position: fixed;
-  top: 0px;
-  left: 0;
-}
-.Continer {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  background-color: #f1f1f1;
-}
-.title {
-  font-size: 24px;
-  font-weight: bold;
-  margin-left: 10px;
-  color: #333;
-  margin: 20px;
-  text-align: center;
-}
-.continer {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 10px;
-  padding: 10px;
-  margin-left: 300px;
-}
-
-.item {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  padding: 10px;
-  border-bottom: 1px solid #ccc;
-}
-.item .imgContiner {
-  position: relative;
-  width: 100%;
-  height: 200px;
-}
-.item:hover {
-  background-color: #ccc;
-  cursor: pointer;
-}
-.item h1 {
-  font-size: 18px;
-  font-weight: bold;
-  margin-left: 10px;
-  color: #333;
-}
-.item .img {
-  width: 100%;
-  height: 150px;
-  object-fit: cover;
-  margin-left: 10px;
-}
-.item p {
-  font-size: 14px;
-  font-weight: normal;
-  margin-left: 10px;
-  color: #333;
-}
