+++ src/app/_Components/Catagory/Catagory.module.css
@@ -1,0 +1,34 @@
+.catagory {
+  display: flex;
+  flex-direction: row;
+  gap: 50px;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  margin: 20px 0;
+  padding: 10px;
+}
+.catagoryItem {
+  text-decoration: none;
+  width: 500px;
+  height: 400px;
+  background-color: #f9f9f9;
+  border-radius: 12px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
+  overflow: hidden;
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  cursor: pointer;
+  text-align: center;
+color: black;
+}
+
+.catagoryItem:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.1);
+}
+
+.imgContiner {
+  position: relative;
+  width: 100%;
+  height: 200px;
+}
