+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Roboto, Mulish } from "next/font/google";
-import Headr from "@/_Components/Header/Headr";
-
+import Headr from "@/src/app/_Components/Header/Headr";
 import "./globals.css";
 
 const Robotofont = Roboto({
@@ -20,8 +19,6 @@ export const metadata: Metadata = {
   title: "News App",
   description: "Generated by create next app",
 };
-
-
 
 export default function RootLayout({
   children,
