+++ src/app/news-list/[category]/page.module.css
@@ -1,0 +1,69 @@
+.mainContinar {
+    position: relative;
+    display: flex;
+    flex-direction: row;
+  }
+  .sidebar {
+    width: 250px;
+    background-color: #3a3939;
+    height:100% ;
+  }
+  .Continer {
+    display: flex;
+    justify-content: flex-start;
+    flex-direction: column;
+    width: 100%;
+    height: 100%;
+    background-color: #f1f1f1;
+  }
+  .title {
+    font-size: 24px;
+    font-weight: bold;
+    margin-left: 10px;
+    color: #333;
+    margin: 20px;
+    z-index: 10;
+
+  }
+  .continer {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-gap: 10px;
+    padding: 10px;
+  }
+  
+  .item {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    padding: 10px;
+    border-bottom: 1px solid #ccc;
+  }
+  .item .imgContiner {
+    position: relative;
+    width: 100%;
+    height: 200px;
+  }
+  .item:hover {
+    background-color: #ccc;
+    cursor: pointer;
+  }
+  .item h1 {
+    font-size: 18px;
+    font-weight: bold;
+    margin-left: 10px;
+    color: #333;
+  }
+  .item .img {
+    width: 100%;
+    height: 150px;
+    object-fit: cover;
+    margin-left: 10px;
+  }
+  .item p {
+    font-size: 14px;
+    font-weight: normal;
+    margin-left: 10px;
+    color: #333;
+  }
