+++ src/app/news-list/[category]/page.tsx
@@ -1,19 +1,19 @@
 
-
 import Image from 'next/image'
-import { fetchData } from '@/services/news.service';
+import { fetchData } from '@/src/app/services/news.service';
 import styles from './page.module.css'
+import Sidebar from './Sidbar';
 interface Iporops {
     params: Promise<{ category: string }>
 }
 const country = "us"
 let category = ''
-
 const Page = async ({ params }: Iporops) => {
     category = (await params).category;
     const latestNews: News.item[] = await fetchData(country, category);
     return (
-        <div>
+        <div className={styles.mainContinar}>
+            <Sidebar />
             <div className={styles.Continer}>
                 <h1 className={styles.title}>{category.toUpperCase() + " "}NEWS</h1>
                 <div className={styles.continer}>
