+++ src/app/page.tsx
@@ -1,0 +1,15 @@
+import Hero from "@/src/app/_Components/Hero/Hero";
+import styles from "./page.module.css";
+import LastNews from "@/src/app/_Components/LastNews/LastNews";
+import Catagory from "@/src/app/_Components/Catagory/Catagory";
+import {Catagorydata} from "@/src/app/data/catagory"
+
+export default function Home() {
+  return (
+    <div className={styles.page}>
+      <Hero />
+      <LastNews />
+      <Catagory data={Catagorydata} />
+    </div>
+  );
+}
